+++ .devcontainer/default-opentofu/devcontainer.json
@@ -1,9 +1,12 @@
 {
-	"name": "Terraform exporter for SAP BTP - Dev (with env file)",
+	"name": "Terraform exporter for SAP BTP - Dev OpenTofu",
+	//https://mcr.microsoft.com/en-us/product/devcontainers/go/tags
 	"image": "mcr.microsoft.com/devcontainers/go:1.23-bullseye",
 	// Features to add to the dev container. More info: https://containers.dev/features.
 	"features": {
-		"ghcr.io/devcontainers/features/terraform:1": {},
+		"ghcr.io/robbert229/devcontainer-features/opentofu:1": {
+			"version": "1.9.0"
+		},
 		"ghcr.io/devcontainers/features/github-cli:1": {}
 	},
 	"customizations": {
@@ -11,24 +14,17 @@
 			"settings": {},
 			"extensions": [
 				"golang.go",
-				"HashiCorp.terraform",
+				"gamunu.opentofu",
 				"ms-vscode.makefile-tools"
 			]
 		},
 		"codespaces": {}
 	},
+	// If you want to use SSO in the devcontainer, you must install the xdg-utils package (see .devcontainer/scripts/install-xdg.sh).
 	"hostRequirements": {
 		"memory": "4gb"
 	},
 	// Use 'forwardPorts' to make a list of ports inside the container available locally.
 	// "forwardPorts": [],
-	"remoteUser": "vscode",
-	// This devcontainer expects a file named .devcontainer/devcontainer.env to exist.
-	// you should place the following environment variables in that file:
-	// - BTP_USERNAME
-	// - BTP_PASSWORD
-	"runArgs": [
-		"--env-file",
-		".devcontainer/devcontainer.env"
-	]
+	"remoteUser": "vscode"
 }
