+++ .devcontainer/withenvfile-opentofu/devcontainer.json
@@ -1,9 +1,11 @@
 {
-	"name": "Terraform exporter for SAP BTP - Dev (with env file)",
+	"name": "Terraform exporter for SAP BTP - Dev OpenTofu (with env file)",
 	"image": "mcr.microsoft.com/devcontainers/go:1.23-bullseye",
 	// Features to add to the dev container. More info: https://containers.dev/features.
 	"features": {
-		"ghcr.io/devcontainers/features/terraform:1": {},
+		"ghcr.io/robbert229/devcontainer-features/opentofu:1": {
+			"version": "1.9.0"
+		},
 		"ghcr.io/devcontainers/features/github-cli:1": {}
 	},
 	"customizations": {
@@ -11,12 +13,13 @@
 			"settings": {},
 			"extensions": [
 				"golang.go",
-				"HashiCorp.terraform",
+        "gamunu.opentofu",
 				"ms-vscode.makefile-tools"
 			]
 		},
 		"codespaces": {}
 	},
+	// If you want to use SSO in the devcontainer, you must install the xdg-utils package (see .devcontainer/scripts/install-xdg.sh).
 	"hostRequirements": {
 		"memory": "4gb"
 	},
